requirements.txt
flask==3.0.0
werkzeug==3.0.1
pdfplumber==0.11.4
pillow==10.4.0
pytesseract==0.3.10
beautifulsoup4==4.12.3
requests==2.32.3
langchain==0.2.14
langchain-community==0.2.12
faiss-cpu==1.8.0
langchain-text-splitters==0.2.2
